jwt-auth-system/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
